Restructure code into mining and analysis sub-packages
